ai-learning-companion/requirements.txt
streamlit>=1.31.0
openai-whisper>=20231117
torch>=2.1.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
langchain>=0.2.11
requests>=2.31.0
numpy>=1.26.0